requirements.txt
pyyaml==6.0
rich==12.6.0
speedtest-cli==2.1.3